requirements.txt
django==3.2.6
django-environ==0.9.0
djangorestframework==3.13.1
psycopg2-binary==2.9.5
djangorestframework-simplejwt==5.2.2
django-phonenumber-field==7.0.2
Pillow==9.2.0
